cleanup generated formats
