--- factom-oapi-generated-clients/haskell-server-generated/factom-open.cabal
@@ -1,33 +1,0 @@
-name:                factom-open
-version:             0.1.0.0
-synopsis:            Auto-generated API bindings for factom-open
-description:         Please see README.md
-homepage:            https://github.com/swagger-api/swagger-codegen#readme
-author:              Author Name Here
-maintainer:          author.name@email.com
-copyright:           YEAR - AUTHOR
-category:            Web
-build-type:          Simple
-cabal-version:       >=1.10
-
-library
-  hs-source-dirs:      lib
-  exposed-modules:     FactomOpen.API
-                     , FactomOpen.Types
-  ghc-options:         -Wall
-  build-depends:       base
-                     , aeson
-                     , text
-                     , containers
-                     , network-uri
-                     , servant
-                     , http-api-data
-                     , servant-client
-                     , servant-server
-                     , servant
-                     , warp
-                     , transformers
-                     , mtl
-                     , http-client
-                     , http-types
-  default-language:    Haskell2010
