+++ factom-oapi-server/stack.yaml.lock
@@ -3,7 +3,21 @@
 # For more information, please see the documentation at:
 #   https://docs.haskellstack.org/en/stable/lock_files
 
-packages: []
+packages:
+- completed:
+    cabal-file:
+      size: 1945
+      sha256: cd005a9231152c94d1443d158d97314ab2948ebdc836acf269cb89c8c7d5a5f1
+    name: api-rpc-factom
+    version: 0.1.3.2
+    git: git@github.com:kompendium-llc/api-rpc-factom
+    pantry-tree:
+      size: 785
+      sha256: 7293225b5069de7410b4ce0cb4f7c33f94a51207c63aed8a872ccd74847a3da5
+    commit: d5cb7930ccefea5717bb13e9834c4af6a24f7be8
+  original:
+    git: git@github.com:kompendium-llc/api-rpc-factom
+    commit: d5cb7930ccefea5717bb13e9834c4af6a24f7be8
 snapshots:
 - completed:
     size: 523884
