+++ factom-oapi-types/api-factom-types.cabal
@@ -19,6 +19,18 @@ library
   build-depends:       base >= 4.7 && < 5
                      , aeson
                      , text
+                     , time
+                     , http-conduit
+                     , servant
+                     , servant-client
+                     , servant-swagger
+                     , http-client
+                     , mtl
+                     , http-types
+                     , http-client-tls
+                     , network
+                     , transformers
+                     , bytestring
                      , containers
   default-language:    Haskell2010
